requirements.txt
fastapi 
uvicorn 
chromadb 
python-dotenv 
langchain-text-splitters
cohere